refactor: abstracts host address resolution and network operations
This commit abstracts address resolution logic and related network
operations behind interfaces. This refactoring simplifies unit testing
by encapsulating side-effect-inducing code behind interfaces,
improving code maintainability and testability.
