--- pkg/ping/interface.go
@@ -1,7 +1,0 @@
-package ping
-
-import "bytes"
-
-type Serializer interface {
-	Serialize() *bytes.Buffer
-}
